Added code to require proof-of-ownership with every transaction. Check is only done when not in develpment node. Whether a check is done, succeeds or fails will be logged.
Background: In order to ensure people transacting actually own the bitcoin address they're transacting with, we'll require them to sign a message with that address' private key. The message will be the transaction ID of the transaction that moved the funds from their wallet to the PZ wallet. They will need to provide this with every payment transaction, and checking this will be part of the blockchain rules, checked by the validators. The client could easily take this once and provide with every transaction.

Signing a message like this is quite standard and supported by the bitcoin wallets. It's pretty common and user-level howtos for how to do it can be found in the web

Added a field to the config file to indicate if we're in development node.
To complete the verification of ownership we'll need to know the TXID of the transaction that deposited the funds to the payment zone, so we can verify the signed message appropriately as detailed above. This will be done once we're integrated with bcoin..